doc:Remove travis badge due to multi-targetting issues
Since changing the app to multi-target both .netcore and netfx, Travis isn't doing test builds cleanly. It's a problem with the build instruction itself, as the build locally is fine, and the Travis output is suggesting I've got something wrong in the yaml. But whilst I'm working that out on another repo as it's going to involve many commits of the yaml it seems, I'm disabling Travis indicators here so it's not misleading.